создала PointController, renderCard теперь метод PointController
